Refactored the data models some by moving properties to more appropriate places. Added some properties that needed to be added. Designed first round (v1) of the TutorialViewController and added a second page to the tutorial process (Tutorial2ViewController).
Note: Should probably convert the tutorial view controllers into a scrolling page view controller. Will need to look up how to do that. Still have not connected any outlets or actions.

Known issues: Tutorial page 2 slides up from the bottom instead of pushing in from the right. I wasn't really wanting to put it in a Navigation Controller, but that may be the simplest and most effective option at this point. Look into the page view controller but don't spend too much time on it. You are on a VERY TIGHT Deadline if you want to get this wrote and submitted with a WatchKit extension for the opening day of Apple Watch sales!